+++ app/uk/gov/hmrc/trustregistration/models/BusinessAsset.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.trustregistration.models
 import play.api.libs.json.Json
 
 case class BusinessAsset(businessDescription: String,
-                         businessValue: Int,
+                         businessValue: Long,
                          company: Company)
 
 object BusinessAsset{
